move mailer mock to conftest
it's simple enough at this point anyway
